# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Urban Arena API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Activity {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  ageLimit: number;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroVideoUrl?: string | null;
  /** @nullable */
  cardImageUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  ctaText: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateActivityBody {
  name: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  ageLimit: number;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroVideoUrl?: string | null;
  /** @nullable */
  cardImageUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isActive?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
  ctaText?: string;
}

export interface UpdateActivityBody {
  name?: string;
  slug?: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  ageLimit?: number;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroVideoUrl?: string | null;
  /** @nullable */
  cardImageUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isActive?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
  ctaText?: string;
}

export interface ReorderActivityBody {
  sortOrder: number;
}

export interface Setting {
  id: number;
  key: string;
  value: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertSettingBody {
  key: string;
  value: string;
}

export interface AdminUser {
  id: number;
  email: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface LoginResponse {
  user: AdminUser;
  token: string;
}

export interface UploadResponse {
  url: string;
  filename: string;
}

export type UploadImageBody = {
  file?: Blob;
};

export type UploadVideoBody = {
  file?: Blob;
};