# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Urban Arena API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all activities
 */
export const ListActivitiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean(),
  isFeatured: zod.boolean(),
  sortOrder: zod.number(),
  ctaText: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListActivitiesResponse = zod.array(ListActivitiesResponseItem);

/**
 * @summary List active activities for public display
 */
export const ListDisplayActivitiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean(),
  isFeatured: zod.boolean(),
  sortOrder: zod.number(),
  ctaText: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListDisplayActivitiesResponse = zod.array(
  ListDisplayActivitiesResponseItem,
);

/**
 * @summary Get activity by ID
 */
export const GetActivityParams = zod.object({
  id: zod.coerce.number(),
});

export const GetActivityResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean(),
  isFeatured: zod.boolean(),
  sortOrder: zod.number(),
  ctaText: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update an activity
 */
export const UpdateActivityParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateActivityBody = zod.object({
  name: zod.string().optional(),
  slug: zod.string().optional(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number().optional(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean().optional(),
  isFeatured: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
  ctaText: zod.string().optional(),
});

export const UpdateActivityResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean(),
  isFeatured: zod.boolean(),
  sortOrder: zod.number(),
  ctaText: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete an activity
 */
export const DeleteActivityParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Create a new activity
 */
export const CreateActivityBody = zod.object({
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean().optional(),
  isFeatured: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
  ctaText: zod.string().optional(),
});

/**
 * @summary Update sort order of an activity
 */
export const ReorderActivityParams = zod.object({
  id: zod.coerce.number(),
});

export const ReorderActivityBody = zod.object({
  sortOrder: zod.number(),
});

export const ReorderActivityResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  shortDescription: zod.string().nullish(),
  fullDescription: zod.string().nullish(),
  ageLimit: zod.number(),
  termsAndConditions: zod.string().nullish(),
  heroImageUrl: zod.string().nullish(),
  heroVideoUrl: zod.string().nullish(),
  cardImageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  isActive: zod.boolean(),
  isFeatured: zod.boolean(),
  sortOrder: zod.number(),
  ctaText: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get all settings
 */
export const ListSettingsResponseItem = zod.object({
  id: zod.number(),
  key: zod.string(),
  value: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListSettingsResponse = zod.array(ListSettingsResponseItem);

/**
 * @summary Upsert a setting
 */
export const UpsertSettingBody = zod.object({
  key: zod.string(),
  value: zod.string(),
});

export const UpsertSettingResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  value: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current admin user
 */
export const GetAuthMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
});

/**
 * @summary Upload an image file
 */
export const UploadImageBody = zod.object({
  file: zod.instanceof(File).optional(),
});

export const UploadImageResponse = zod.object({
  url: zod.string(),
  filename: zod.string(),
});

/**
 * @summary Upload a video file
 */
export const UploadVideoBody = zod.object({
  file: zod.instanceof(File).optional(),
});

export const UploadVideoResponse = zod.object({
  url: zod.string(),
  filename: zod.string(),
});