# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Urban Arena API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  ageLimit: number;
  /** @nullable */
  termsAndConditions?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  heroVideoUrl?: string | null;
  /** @nullable */
  cardImageUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  ctaText: string;
  createdAt: Date;
  updatedAt: Date;
}