# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Urban Arena API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  email: string;
}