# lib/api-zod/src/generated/types/upsertSettingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Urban Arena API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertSettingBody {
  key: string;
  value: string;
}